triton_mla_sparse: promote KV row offsets to int64; merge upstream hardening
- The sparse-MLA Triton kernel computed `indices * stride_kv_token` in int32.
  A row id fits in int32 but the element offset does not once the flat KV
  view exceeds ~4.2M rows (1M-context configs reach this). Promote the row
  term to int64 in all three loads (K, K_pe, V), mirroring what upstream's
  _sparse_kv_row_offset does on the ROCm path. Not the cause of the current
  cudagraph fault (today's sizes stay under the limit) -- a long-context
  correctness fix found while auditing for the same class of bug upstream
  just fixed in the GDN prefill path (cu_seqlens -> int64).

- Also merges upstream PR #53906 commit 878631b60 ("Harden GLM-5.3
  FlashInfer integration"): kernel_block_size becomes an explicit builder
  interface on the base class, GDN cu_seqlens normalized to int64, config
  validation for unsupported GLM variants.

Debug instrumentation (env-gated VLLM_GLM5_KPOOL_DEBUG bounds checks in the
kpool/KDA/sparse-MLA paths) is retained for now; it will be stripped before
the branch is finalized.

+++ vllm/v1/attention/ops/triton_mla_sparse_kernel.py
@@ -119,7 +119,7 @@ def _sparse_mla_compute_tile(
         mask_kv = (indices >= 0) & (indices < seq_kv)
 
         offs_k = (
-            indices[None, :] * stride_kv_token
+            indices[None, :].to(tl.int64) * stride_kv_token
             + cur_kv_head_id * stride_kv_head
             + offs_d[:, None]
         )
@@ -128,7 +128,7 @@ def _sparse_mla_compute_tile(
 
         if BLOCK_DPE > 0:
             offs_kpe = (
-                indices[None, :] * stride_kv_token
+                indices[None, :].to(tl.int64) * stride_kv_token
                 + cur_kv_head_id * stride_kv_head
                 + offs_dpe[:, None]
             )
@@ -143,7 +143,7 @@ def _sparse_mla_compute_tile(
         qk = tl.where((mask_h[:, None]) & (mask_kv[None, :]), qk, NEG_LARGE)
 
         offs_v = (
-            indices[:, None] * stride_kv_token
+            indices[:, None].to(tl.int64) * stride_kv_token
             + cur_kv_head_id * stride_kv_head
             + offs_dv[None, :]
         )
